requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.7.0
numpy>=1.24.0
Pillow>=9.5.0
segment-anything-hq>=0.3.0
transformers>=4.35.0
einops
timm  
webcolors
elevenlabs
python-dotenv